AniGen/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
astor=0.8.1=pypi_0
bashutils=0.0.4=pypi_0
ca-certificates=2021.1.19=h06a4308_0
cached-property=1.5.2=pypi_0
certifi=2020.12.5=py37h06a4308_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
easydict=1.9=pypi_0
filelock=3.0.12=pypi_0
fire=0.4.0=pypi_0
flask=1.1.2=pypi_0
gast=0.2.2=pypi_0
google-pasta=0.2.0=pypi_0
gpt-2-simple=0.7.2=pypi_0
grpcio=1.35.0=pypi_0
h5py=3.1.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=3.4.0=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.0.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
layers=0.1.5=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.3=pypi_0
markupsafe=1.1.1=pypi_0
ncurses=6.2=he6710b0_1
nltk=3.5=pypi_0
numpy=1.20.1=pypi_0
openssl=1.1.1j=h27cfd23_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pypi_0
paho-mqtt=1.5.1=pypi_0
pillow=8.1.0=pypi_0
pip=21.0.1=py37h06a4308_0
protobuf=3.15.0=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.9=h7579374_0
pytorch-model-summary=0.1.2=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
regex=2020.11.13=pypi_0
requests=2.25.1=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.1=pypi_0
sentence-transformers=0.4.1.2=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=53.0.0=pypi_0
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
tensorflow-gpu=1.15.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.1=pypi_0
toposort=1.6=pypi_0
torch=1.7.1=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.8.2=pypi_0
tqdm=4.57.0=pypi_0
transformers=4.3.2=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.3=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
